django-taxi/requirements.txt
channels==2.1.2
channels-redis==2.3.0
Django==2.1
djangorestframework==3.8.2
nose==1.3.7
Pillow==5.2.0
pytest-asyncio==0.9.0 
pytest-django==3.4.2